API key authentication works in eg-api.php, in a basic way.
A user submits a request with the parameters 'apikey' and 'useremail'.
Then eg-api queries the database for the hashed api key associated with
the given useremail. eg-api uses PHP's built-in password_verify
function to check if the submitted api key matches the hashed api key
that has been stored.

For this to work, the apikey has to be created in a particular way.

First, there must be a table column in the table user called apikey.

Then an api key must be set using PHP's built in
password_hash([password],[hashfunction]). password_hash is necessary
so that verify_password can work.

How the API key itself is created is not important for the rest of the
system, but I did it for testing purposes with the following function
that relied on php's uniqid function, with the EXTRA_ENTROPY variable
set to True:

        function generateAPIKey($useremail) {
	                return base64_encode(uniqid($useremail, True));
	}